+++ source/Grid/COME_AbsTopologicalComponent.hpp
@@ -1,9 +1,12 @@
 #pragma once
 #include <iostream>
 #include <vector>
-#include "COME_Element.hpp"
+#include <unordered_map>
 namespace Mesh
 {
+	template <int dim, int spacedim>
+	class Element;
+
 	template <int dim, int spacedim = dim>
 	class AbsTopologicalComponent
 	{
@@ -15,8 +18,22 @@ namespace Mesh
 		const std::vector<Element<dim, spacedim>*>& getAdjacentElements() const;
 
 	protected:
-		unsigned int polynomialDegree_;
+		unsigned int polynomialDegree_ = 1;
 		bool isActive_ = false;
 		std::vector<Element<dim,spacedim>*> listOfAdjacentElements_;
+		unsigned int refinementLevel_ = 0;
+
 	};
+
+	template<int dim, int spacedim>
+	void AbsTopologicalComponent<dim, spacedim>::addAdjacentElement(Element<dim, spacedim>* el)
+	{
+		listOfAdjacentElements_.push_back(el);
+	}
+	template<int dim, int spacedim>
+	const std::vector<Element<dim, spacedim>*>& AbsTopologicalComponent<dim, spacedim>::getAdjacentElements() const
+	{
+		return listOfAdjacentElements_;
+	}
+
 }