+++ source/Grid/COME_Mesh.hpp
@@ -3,6 +3,7 @@
 #include "COME_Edge.hpp"
 #include "COME_Element.hpp"
 #include "COME_Face.hpp"
+#include "COME_Volume.hpp"
 #include "Grid/COME_AbsTopologicalComponent.hpp"
 
 #include <iostream>
@@ -14,6 +15,7 @@
 #include <sstream>
 #include <array>
 #include <unordered_map>
+#include <map>
 
 namespace Mesh
 {
@@ -30,9 +32,10 @@ namespace Mesh
 		Mesh() = default;
 		Mesh(const Mesh&) = delete;
 		Mesh& operator=(const Mesh&) = delete;
-		void addEdge();
+		void addEdge(int node1, int node2);
 		void addNode(int nodeindex,const std::array<double,spacedim>& positions);
 		void addFace();
+		void addVolume();
 		void addElement(int elementnumber, const std::array<int, 1 << spacedim>& nodes);
 
 		void read_abaqus(std::istream& in);
@@ -42,9 +45,13 @@ namespace Mesh
 		std::vector<std::unique_ptr<Element<dim,spacedim>>>		listOfElements_;
 		std::vector<std::unique_ptr<Face<dim,spacedim>>>		listOfFaces_;
 		std::vector<std::unique_ptr<Node<dim,spacedim>>>		listOfNodes_;
+		std::vector<std::unique_ptr<Volume<dim, spacedim>>>		listOfVolumes_;
 
 		std::vector<std::string> split_csv(const std::string& line) const;
+
 		std::unordered_map<int, Node<dim, spacedim>*> nodeIdMap_;
+		std::map<std::pair<int, int>, Edge<dim, spacedim>*> edgeMap_; //smart way of storing an Edge pointer with a node mapping.
+		std::map<std::array<int, 4>, Face<dim, spacedim>*> faceMap_;